Refactor method for receiving feedback.
We decided on a new format when sending the feedback from the frontend to the backend.
This required changing the already existing method calls and adding some new ones for open feedback and feedback questions.
This new functionality relates to issues #13 and #20, and the rest of the code for this is implemented there.