Export batch 2 match data to CSV file.
Added functionality in DashboardController for exporting a CSV file with the feedback provided by each user about the recommendations he was matched into.
Each row will consist of userId, and for each match, the id of the matched user, the rating for the playlist he was recommended, the feedback he provided for each song and open feedback if there was any.
The SQL functionality for such a request has been added in DashboardService, and it returns a big list of tuples with all the information together.