Export batch 2 song rating data to CSV file.
Added functionality in DashboardController for exporting a CSV file with the feedback provided by each user about each song he was recommended to.
Each row will consist of userId, and 15 fields representing the 15 songs the user will have been recommended. In each field there will be the rating the user gave it, or a 0 if there was no rating provided.
The SQL functionality for such a request has been added in DashboardService.get_song_ratings, and it returns one row per user with a concatenation of the songs and its position in the playlist.
The formatting of the final CSV row is done in DashboardController.retrieve_song_ratings.