Added SongRating field.
Created a new field playlistNumber to each song rating to keep track of the order of it in the database when stored. This new field acts as [1-5] index indicating which song belongs to which position in someone's ratings. IT's also updated in the database and it helps with the creation of the CSV file.