Fix bug in match data retrieval.
Last field in the CSV file for the match and feedback information (random_open_feedback) was shown letter by letter instead of in one block of text. This was fixed by changing one line of code.

Additional fix: based on our client's suggestions, we changed the calculation method of the personality scores in psychology.py. This way, we normalize the scores by substracting each value by the total mean computed across all retrieved answers. The test for this functionality has been updated as well.