Update service and controller to the latest convention.
Service changes:
We created a new database schema so we had to update the SQL commands in the database to match the names there. Naming convention is CamelCase for table names and database names and camelCase for fields.
The SQL commands were also updated to have consistent syntax in the main words.
Some of the methods in the services didn't apply the commit rule properly (with the env file) so that was added as well.
add_song_ratings() in SongService.py hasn't been updated because it is an outdated method that will get deleted later on.

Controller changes:
Uncommented some methods leftover from previous sprint which had to be done after merging everything.
We didn't store personality or value scores at any point in the code so that has been added as well.