Update SQL commands and refactor database opening.
Database connection is now created in the DashboardController methods and it is passed along in the service method calls.
Database name is now also abstracted and passed as a variable in the SQL queries.
These queries also have been modified to have their keywords in capital letters according to conventions and to ease the reading of such big commands.