Update methods to GET.
Some methods were supposed to be GET methods but we made them POST due to the fact that we used a body for the data. The data is now provided in the URL as fields and they are updated to GET methods now.
Additional change: added table name to ambiguous SQL query.