Add experiment parameter retrieval and modification.
The frontend can now access batch and metric information, as well as change this parameters and put the experiment into batch 2. Additionally, another endpoint has been added to retrieve the total number of users that have completed the questionnaire, and the usage of batch numbers across the code has been abstracted to retrieve the number from the backend .env file.