Implement wait for backend to return results before showing graph
After the user is done with answering questions, he/she will be redirected to the results page. After redirection, the sendAnswer question will be triggered, which will send all the questionnaire answers to the back-end, which will send back the personality and values results. Once received they will be shown with 2 graphs.